requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.25.0
httpx>=0.27.0
redis>=5.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
